#1334 cleanup range and colot
