+++ lib/elementWrapper/elementWrapperFactory.js
@@ -26,10 +26,10 @@ const createElementWrapper = (elementType, query, attrValuePairs, _options, ...a
       return new CheckBoxWrapper(selector, options, description);
     case 'DropDown':
       return new DropDownWrapper(selector, options, description);
-    // case 'colorField':
-    //   return new ColorWrapper(get, description, getIfExists);
-    // case 'rangeType':
-    //   return new RangeWrapper(get, description, getIfExists);
+    case 'ColorField':
+      return new ColorWrapper(selector, options, description);
+    case 'RangeType':
+      return new RangeWrapper(selector, options, description);
     // case 'timeField':
     //   return new TimeFieldWrapper(get, description, getIfExists);
     // default:
