+++ lib/elementWrapper/rangeWrapper.js
@@ -1,8 +1,51 @@
 const Range = require('../elements/range');
 const ValueWrapper = require('./valueWrapper');
-const { firstElement } = require('./helper');
+const { firstElement, getElementGetter } = require('./helper');
+const { $function } = require('../elementSearch');
 
+function getRangeElementWithLabel(label) {
+  const fileField = [];
+  function checkAndPushElement(elem) {
+    if (
+      elem.tagName &&
+      elem.tagName.toLowerCase() == 'input' &&
+      elem.type &&
+      elem.type.toLowerCase() === 'range'
+    ) {
+      fileField.push(elem);
+    }
+  }
+  const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
+    return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
+  });
+  for (let matchingLabel of matchingLabels) {
+    const labelFor = matchingLabel.getAttribute('for');
+    if (labelFor) {
+      //check label with attribute for
+      const labelForElement = document.getElementById(labelFor);
+      checkAndPushElement(labelForElement);
+    } else {
+      // check child node of label tag
+      matchingLabel.childNodes.forEach((elem) => {
+        checkAndPushElement(elem);
+      });
+    }
+  }
+  return fileField;
+}
 class RangeWrapper extends ValueWrapper {
+  _get = getElementGetter(
+    this.selector,
+    async () =>
+      await $function(
+        getRangeElementWithLabel,
+        this.selector.label,
+        this.options.selectHiddenElements,
+      ),
+    'input[type="range"]',
+    this.options.selectHiddenElements,
+  );
+
   async select(value) {
     const elem = await firstElement.apply(this);
     return await elem.select(value);
