+++ lib/taiko.js
@@ -2302,47 +2302,7 @@ module.exports.range = range;
 
 function range(attrValuePairs, _options = {}, ...args) {
   validate();
-  const { selector, options } = prepareParameters(attrValuePairs, _options, ...args);
-  function getRangeElementWithLabel(label) {
-    const fileField = [];
-    function checkAndPushElement(elem) {
-      if (
-        elem.tagName &&
-        elem.tagName.toLowerCase() == 'input' &&
-        elem.type &&
-        elem.type.toLowerCase() === 'range'
-      ) {
-        fileField.push(elem);
-      }
-    }
-    const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
-      return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
-    });
-    for (let matchingLabel of matchingLabels) {
-      const labelFor = matchingLabel.getAttribute('for');
-      if (labelFor) {
-        //check label with attribute for
-        const labelForElement = document.getElementById(labelFor);
-        checkAndPushElement(labelForElement);
-      } else {
-        // check child node of label tag
-        matchingLabel.childNodes.forEach((elem) => {
-          checkAndPushElement(elem);
-        });
-      }
-    }
-    return fileField;
-  }
-  const get = getElementGetter(
-    selector,
-    async () =>
-      await $function(getRangeElementWithLabel, selector.label, options.selectHiddenElements),
-    'input[type="range"]',
-    options.selectHiddenElements,
-  );
-
-  const description = desc(selector, 'label', 'Range Type');
-  return createElementWrapper(get, description, 'rangeType');
+  return createElementWrapper('RangeType', 'label', attrValuePairs, _options, ...args);
 }
 
 /**
@@ -2369,47 +2329,8 @@ function color(attrValuePairs, _options = {}, ...args) {
   if (attrValuePairs === undefined || attrValuePairs === '') {
     throw new Error('Atleast one attribute is require!');
   }
-  const { selector, options } = prepareParameters(attrValuePairs, _options, ...args);
-  function getColorElementWithLabel(label) {
-    const fileField = [];
-    function checkAndPushElement(elem) {
-      if (
-        elem.tagName &&
-        elem.tagName.toLowerCase() == 'input' &&
-        elem.type &&
-        elem.type.toLowerCase() === 'color'
-      ) {
-        fileField.push(elem);
-      }
-    }
-    const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
-      return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
-    });
-    for (let matchingLabel of matchingLabels) {
-      const labelFor = matchingLabel.getAttribute('for');
-      if (labelFor) {
-        //check label with attribute for
-        const labelForElement = document.getElementById(labelFor);
-        checkAndPushElement(labelForElement);
-      } else {
-        // check child node of label tag
-        matchingLabel.childNodes.forEach((elem) => {
-          checkAndPushElement(elem);
-        });
-      }
-    }
-    return fileField;
-  }
-  const get = getElementGetter(
-    selector,
-    async () =>
-      await $function(getColorElementWithLabel, selector.label, options.selectHiddenElements),
-    'input[type="color"]',
-    options.selectHiddenElements,
-  );
 
-  const description = desc(selector, 'label', 'Color field');
-  return createElementWrapper(get, description, 'colorField');
+  return createElementWrapper('ColorField', 'label', attrValuePairs, _options, ...args);
 }
 
 /**
